pass rCore-ch2, but need "-bios rustsbi-qemu.bin"

+++ os/src/console.rs
@@ -8,7 +8,7 @@ struct Stdout;
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            console_putchar(c as u8);
+            console_putchar(c as usize);
         }
         Ok(())
     }
